pyproject.toml
[tool.poetry]
name = "poopy"
version = "0.1.0"
description = ""
authors = ["syafiq <ahmadsyafiq93@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
